+++ templates/hexarch/Domain/Domain.Data/AssemblyInfo.cs
@@ -1,3 +1,6 @@
 namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Data;
 
+/// <summary>
+/// Marker interface for assembly information.
+/// </summary>
 public interface IAssemblyInfo { }