+++ templates/hexarch/Domain/Domain.Logic.Test/SakServiceTests.cs
@@ -12,15 +12,12 @@ namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Logic.Test;
 
 public class SakServiceTests
 {
-    private readonly SakService _sut;
+    private readonly SakService _sut; // System Under Test
     private readonly ISakRepository _sakRepositoryMock = Substitute.For<ISakRepository>();
 
-    private static readonly string SampleOrgNr = "123456789";
+    private const string SampleOrgNr = "123456789";
 
-    private readonly DomainConfiguration _domainConfiguration = new()
-    {
-        SomeSetting = "SampleConfigValue",
-    };
+    private readonly DomainConfiguration _domainConfiguration = new() { SakDeadlineDays = 30 };
 
     public SakServiceTests()
     {
@@ -33,11 +30,13 @@ public class SakServiceTests
         //arrange
         var mockedSakResponse = new Faker<Sak>().Generate() with
         {
-            Organisajonsnummer = SampleOrgNr,
+            Organisasjonsnummer = SampleOrgNr,
         };
-        _sakRepositoryMock.PersistSak(SampleOrgNr).Returns(mockedSakResponse);
+        _sakRepositoryMock.PersistSak(default!).ReturnsForAnyArgs(mockedSakResponse);
         //act
-        var result = await _sut.CreateNewSak(new CreateSakDto { Organisajonsnummer = SampleOrgNr });
+        var result = await _sut.CreateNewSak(
+            new CreateSakDto { Organisasjonsnummer = SampleOrgNr }
+        );
         //assert
         result.ShouldBeEquivalentTo(mockedSakResponse);
     }
@@ -98,8 +97,7 @@ public class SakServiceTests
     {
         //arrange
         var testId = Guid.NewGuid();
-        Sak? mockedSakResponse = null;
-        _sakRepositoryMock.UpdateSakStatus(testId, SakStatus.InProgress).Returns(mockedSakResponse);
+        _sakRepositoryMock.UpdateSakStatus(testId, SakStatus.InProgress).Returns(default(Sak?));
         //act
         var act = () => _sut.StartSak(testId);
         //assert