+++ templates/hexarch/Domain/Domain.Logic/DependencyInjection.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.Reflection;
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.API.Ports;
+using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Data;
 using Mapster;
 using MapsterMapper;
 using Microsoft.Extensions.DependencyInjection;
@@ -9,47 +10,36 @@ using Microsoft.Extensions.Options;
 // ReSharper disable once CheckNamespace
 namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Logic.DependencyInjection;
 
+/// <summary>
+/// Domain specific configuration, e.g. business rules, limits etc.
+/// </summary>
 public class DomainConfiguration
 {
+    /// <summary>
+    /// The length of time before a <see cref="Sak"/> deadline is reached.
+    /// </summary>
     [Required]
-    public required string SomeSetting { get; init; }
+    public required int SakDeadlineDays { get; init; }
 }
 
+/// <summary>
+/// Extension methods to add domain services to the DI container.
+/// </summary>
 public static class DependencyInjection
 {
+    /// <summary>
+    /// Adds domain functionality to the application. This usually includes domain logic, and sometimes configuration.
+    /// </summary>
+    /// <param name="services"></param>
+    /// <param name="domainConfiguration"></param>
+    /// <returns></returns>
     public static IServiceCollection AddDomain(
         this IServiceCollection services,
         DomainConfiguration domainConfiguration
     )
     {
-        services.AddMapper();
         services.AddSingleton(Options.Create(domainConfiguration));
         services.AddScoped<ISakService, SakService>();
-        return services;
-    }
-
-    internal static IServiceCollection AddMapper(this IServiceCollection services)
-    {
-        var existingConfig = services
-            .Select(s => s.ImplementationInstance)
-            .OfType<TypeAdapterConfig>()
-            .FirstOrDefault();
-
-        if (existingConfig == null)
-        {
-            var config = new TypeAdapterConfig()
-            {
-                RequireExplicitMapping = false,
-                RequireDestinationMemberSource = true,
-            };
-            config.Scan(Assembly.GetExecutingAssembly());
-            services.AddSingleton(config);
-            services.AddScoped<IMapper, ServiceMapper>();
-        }
-        else
-        {
-            existingConfig.Scan(Assembly.GetExecutingAssembly());
-        }
         return services;
     }
 }