+++ templates/hexarch/Infrastructure/Infrastructure.Adapters.Test/SakRepositoryTests.cs
@@ -1,48 +1,60 @@
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Domain.Data;
-using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.Db.Model;
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.Test.Fixtures;
 using Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Ports;
-using Bogus;
 using Shouldly;
 using Xunit.Microsoft.DependencyInjection.Abstracts;
 
 namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Adapters.Test;
 
-public class SakRepositoryTests(
-    ITestOutputHelper testOutputHelper,
-    InfrastructureAdapterTestFixture infrastractureAdapterTestFixture
-) : TestBed<InfrastructureAdapterTestFixture>(testOutputHelper, infrastractureAdapterTestFixture)
+public class SakRepositoryTests : TestBed<InfrastructureAdapterTestFixture>
 {
-    private readonly ISakRepository _sut =
-        infrastractureAdapterTestFixture.GetService<ISakRepository>(testOutputHelper)!;
+    private readonly ISakRepository _sut; // System Under Test
 
     private static readonly string SampleOrgNr = "123456789";
+
+    private readonly VerifySettings _verifierSettings = new();
+
+    public SakRepositoryTests(
+        ITestOutputHelper testOutputHelper,
+        InfrastructureAdapterTestFixture infrastractureAdapterTestFixture
+    )
+        : base(testOutputHelper, infrastractureAdapterTestFixture)
+    {
+        _sut = infrastractureAdapterTestFixture.GetService<ISakRepository>(testOutputHelper)!;
+
+        _verifierSettings.DontScrubGuids();
+        _verifierSettings.UseDirectory("Snapshots");
+    }
 
     [Fact]
     public async Task CreateSak_WhenCalled_PersistsSakEntityAsync()
     {
+        // arrange
+        var newSak = TestData.CreateSakFaker(1).Generate() with
+        {
+            Organisasjonsnummer = SampleOrgNr,
+        };
+
         // act
-        var createdSak = await _sut.PersistSak(SampleOrgNr);
+        var createdSak = await _sut.PersistSak(newSak);
         // assert
         var result = await _sut.GetSak(createdSak.Id);
-        result?.Organisajonsnummer.ShouldBe(SampleOrgNr);
+        await Verify(result, _verifierSettings);
     }
 
     [Fact]
     public async Task UpdateSakStatus_WhenCalled_PersistsSakEntityAsync()
     {
         // arrange
-        var createdSak = await _sut.PersistSak(SampleOrgNr);
+        var createdSak = TestData.CreateSakFaker(2).Generate() with
+        {
+            Organisasjonsnummer = SampleOrgNr,
+        };
+        await _sut.PersistSak(createdSak);
         // act
         var updatedSak = await _sut.UpdateSakStatus(createdSak.Id, SakStatus.InProgress);
         // assert
-        updatedSak.ShouldBeEquivalentTo(
-            createdSak with
-            {
-                Status = SakStatus.InProgress,
-                LastUpdated = updatedSak!.LastUpdated,
-            }
-        );
+        await Verify(updatedSak, _verifierSettings);
     }
 
     [Fact]