+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.3.0
+
+### Changed
+
+- added(hexarch): added more doc strings to the public classes
+- added(hexarch): added deadline to the sak (example domain)
+- changed(hexarch): fixed typo in Sak.Organisa**s**jonsnummer
+
 ## 2.2.5
 
 ### Changed